+++ tests/integration/test_comments.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from bigxml import XMLHandler, parse, xml_handle_element
+from bigxml import Parser, XMLHandler, xml_handle_element
 
 
 def test_maths_eval():
@@ -11,8 +11,8 @@ def test_maths_eval():
         def handle_author(self, node):
             yield node.text
 
-    with (Path(__file__).parent / "comments.xml").open("rb") as f_in:
-        items = parse(f_in, Handler())
+    with (Path(__file__).parent / "comments.xml").open("rb") as stream:
+        items = Parser(stream).iter_from(Handler())
         assert next(items) == "Test"
         assert next(items) == "Hello everyone!"
         assert next(items) == (